Fix insane util situation (WIP)
